Split up Solaris and illumos targets
